+++ src/test/scratch.ts
@@ -1,6 +1,6 @@
 import { TaskPaperNode } from "task-parser/TaskPaperNode";
 import { todayDay } from "../dates";
 
-const today = todayDay.format("YYYY-MM-DD");
+const today = todayDay().format("YYYY-MM-DD");
 const testTask = new TaskPaperNode(`  - test item @due(${today}`);